chore: Move CreateMissingOptions to shared Actions\Options namespace - Fix migrations - Allow min:1 for project skills
