+++ app/Actions/Options/CreateMissingOptions.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Actions\Onboarding;
+namespace App\Actions\Options;
 
 use App\Models\Option;
 use App\Models\Category;
@@ -12,7 +12,7 @@ class CreateMissingOptions
     {
         foreach (['domain', 'language', 'framework', 'specialization'] as $categoryName) {
             if ($category = Category::where('name', $categoryName)->first()) {
-                collect($data['skills'][$categoryName])->each(
+                collect($data['skills'][$categoryName] ?? [])->each(
                     fn($optionName) =>
                     Option::firstOrCreate([
                         'name' => strtolower($optionName),
